pyproject.toml
[tool.poetry]
name = "pkgroup-cadsystem"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fast-bitrix24 = "^1.7.1"
flask = "^3.0.3"
flask-restx = "^1.3.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
